% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_operations.R
\name{assays<-,SummarizedExperimentLogged,SimpleList-method}
\alias{assays<-,SummarizedExperimentLogged,SimpleList-method}
\alias{assays<-,SingleCellExperimentLogged,SimpleList-method}
\title{Assign List to assays using \verb{assays<-} for ExperimentLogged}
\usage{
\S4method{assays}{SummarizedExperimentLogged,SimpleList}(x, withDimnames = TRUE, ...) <- value

\S4method{assays}{SingleCellExperimentLogged,SimpleList}(x, withDimnames = TRUE, ...) <- value
}
\description{
Assign List to assays using \verb{assays<-} for ExperimentLogged
}
