% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_operations.R
\name{colData<-,ExperimentLogged,DataFrame-method}
\alias{colData<-,ExperimentLogged,DataFrame-method}
\alias{colData<-,ExperimentLogged,DFrame-method}
\title{Assign column to colData using \verb{colData<-} for ExperimentLogged}
\usage{
\S4method{colData}{ExperimentLogged,DataFrame}(x) <- value

\S4method{colData}{ExperimentLogged,DFrame}(x) <- value
}
\description{
Assign column to colData using \verb{colData<-} for ExperimentLogged
}
