% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_operations.R
\name{$<-,SummarizedExperimentLogged-method}
\alias{$<-,SummarizedExperimentLogged-method}
\alias{$<-,SingleCellExperimentLogged-method}
\title{Assign column to colData using \verb{$<-} for ExperimentLogged}
\usage{
\S4method{$}{SummarizedExperimentLogged}(x, name) <- value

\S4method{$}{SingleCellExperimentLogged}(x, name) <- value
}
\description{
Assign column to colData using \verb{$<-} for ExperimentLogged
}
\examples{
# Create a logged SummarizedExperiment
if (requireNamespace("tidySummarizedExperiment", quietly = TRUE)) {
  se <- tidySummarizedExperiment::pasilla
  se_logged <- log_start(se)
  
  # Modify existing column
  colData(se_logged)$condition <- tolower(colData(se_logged)$condition)
  
  # Add new column
  colData(se_logged)$new_column <- rep("test", ncol(se_logged))
}
}
