% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ExperimentLogged-class}
\alias{ExperimentLogged-class}
\title{ExperimentLogged class}
\description{
A class union of SummarizedExperimentLogged and SingleCellExperimentLogged,
used to define logging methods once for both classes.
}
