% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{extract.SummarizedExperimentLogged}
\alias{extract.SummarizedExperimentLogged}
\alias{extract.SingleCellExperimentLogged}
\title{Extract values from a column into multiple columns in An ExperimentLogged object}
\usage{
\method{extract}{SummarizedExperimentLogged}(
  data,
  col,
  into,
  regex = "([[:alnum:]]+)",
  remove = TRUE,
  convert = FALSE,
  ...
)

\method{extract}{SingleCellExperimentLogged}(
  data,
  col,
  into,
  regex = "([[:alnum:]]+)",
  remove = TRUE,
  convert = FALSE,
  ...
)
}
\arguments{
\item{data}{An ExperimentLogged object}

\item{col}{Column to extract from}

\item{into}{Names of new variables to create}

\item{regex}{A regular expression to extract values}

\item{remove}{If TRUE, remove input column from output}

\item{convert}{If TRUE, runs type.convert() on each new column}

\item{...}{Additional arguments passed to tidyr::extract}
}
\description{
Extract values from a column into multiple columns in An ExperimentLogged object
}
\examples{
if (requireNamespace("tidySummarizedExperiment", quietly = TRUE)) {
  se <- tidySummarizedExperiment::pasilla
  se_logged <- log_start(se)
  result <- se_logged |>
    mutate(type = sub("_end","",colData(se)$type)) |>
    extract(type, into = "exp")
}
}
