% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{filter.SummarizedExperimentLogged}
\alias{filter.SummarizedExperimentLogged}
\alias{filter.SingleCellExperimentLogged}
\title{Filter rows and columns of An ExperimentLogged object}
\usage{
\method{filter}{SummarizedExperimentLogged}(.data, ...)

\method{filter}{SingleCellExperimentLogged}(.data, ...)
}
\arguments{
\item{.data}{An ExperimentLogged object}

\item{...}{Logical expressions used for filtering}
}
\description{
Filter rows and columns of An ExperimentLogged object
}
\examples{
if (requireNamespace("tidySummarizedExperiment", quietly = TRUE)) {
  se <- tidySummarizedExperiment::pasilla
  se_logged <- log_start(se)
  result <- se_logged |>
    filter(condition == "treated")
}
}
