% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_start}
\alias{log_start}
\alias{show,SummarizedExperimentLogged-method}
\alias{show,SingleCellExperimentLogged-method}
\title{Create a logging-enabled SummarizedExperiment object}
\usage{
log_start(se)

\S4method{show}{SummarizedExperimentLogged}(object)

\S4method{show}{SingleCellExperimentLogged}(object)
}
\arguments{
\item{se}{A SummarizedExperiment or derived object}

\item{object}{An ExperimentLogged object}
}
\value{
An ExperimentLogged object with tracking capabilities
}
\description{
This function wraps a SummarizedExperiment object with logging capabilities.
Operations performed on the resulting object will be tracked and displayed
when the object is printed.
}
\examples{
if (requireNamespace("tidySummarizedExperiment", quietly = TRUE)) {
  se <- tidySummarizedExperiment::pasilla
  se_logged <- log_start(se)
  result <- se_logged |>
    filter(condition == "treated")
}
}
