% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{mutate.SummarizedExperimentLogged}
\alias{mutate.SummarizedExperimentLogged}
\alias{mutate.SingleCellExperimentLogged}
\title{Modify columns of An ExperimentLogged object}
\usage{
\method{mutate}{SummarizedExperimentLogged}(.data, ...)

\method{mutate}{SingleCellExperimentLogged}(.data, ...)
}
\arguments{
\item{.data}{An ExperimentLogged object}

\item{...}{Name-value pairs of expressions used to modify columns}
}
\description{
Modify columns of An ExperimentLogged object
}
\examples{
if (requireNamespace("tidySummarizedExperiment", quietly = TRUE)) {
  se <- tidySummarizedExperiment::pasilla
  se_logged <- log_start(se)
  result <- se_logged |>
    mutate(type = sub("_end","",colData(se)$type))
}
}
