% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_operations.R
\name{reducedDim<-,SingleCellExperimentLogged,character-method}
\alias{reducedDim<-,SingleCellExperimentLogged,character-method}
\alias{reducedDim<-,SingleCellExperimentLogged,numeric-method}
\title{Assign a matrix to a named reduced dimension using \verb{reducedDim<-} for SingleCellExperimentLogged}
\usage{
\S4method{reducedDim}{SingleCellExperimentLogged,character}(x, type, withDimnames = TRUE, ...) <- value

\S4method{reducedDim}{SingleCellExperimentLogged,numeric}(x, type, withDimnames = TRUE, ...) <- value
}
\description{
Assign a matrix to a named reduced dimension using \verb{reducedDim<-} for SingleCellExperimentLogged
}
