% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_operations.R
\name{reducedDims<-,SingleCellExperimentLogged-method}
\alias{reducedDims<-,SingleCellExperimentLogged-method}
\title{Assign a List of matrices to reducedDims using \verb{reducedDims<-} for SingleCellExperimentLogged}
\usage{
\S4method{reducedDims}{SingleCellExperimentLogged}(x, withDimnames = TRUE, ...) <- value
}
\description{
Assign a List of matrices to reducedDims using \verb{reducedDims<-} for SingleCellExperimentLogged
}
