% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{mutate}
\alias{select}
\alias{slice}
\alias{extract}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:filter]{filter()}}, \code{\link[dplyr:mutate]{mutate()}}, \code{\link[dplyr:select]{select()}}, \code{\link[dplyr:slice]{slice()}}}

  \item{tidyr}{\code{\link[tidyr:extract]{extract()}}}
}}

