% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_operations.R
\name{rowData<-,ExperimentLogged,DataFrame-method}
\alias{rowData<-,ExperimentLogged,DataFrame-method}
\alias{rowData<-,ExperimentLogged,DFrame-method}
\title{Assign column to rowData using \verb{rowData<-} for ExperimentLogged}
\usage{
\S4method{rowData}{ExperimentLogged,DataFrame}(x) <- value

\S4method{rowData}{ExperimentLogged,DFrame}(x) <- value
}
\description{
Assign column to rowData using \verb{rowData<-} for ExperimentLogged
}
