% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SingleCellExperimentLogged-class}
\alias{SingleCellExperimentLogged-class}
\title{SingleCellExperimentLogged class}
\description{
A class extending SingleCellExperiment to include logging capabilities.
This class tracks operations performed on the object and displays a
log when the object is printed.
}
\section{Slots}{

\describe{
\item{\code{log_history}}{A tibble storing the history of operations}
}}

