% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{slice.SummarizedExperimentLogged}
\alias{slice.SummarizedExperimentLogged}
\alias{slice.SingleCellExperimentLogged}
\title{Slice rows from a ExperimentLogged object}
\usage{
\method{slice}{SummarizedExperimentLogged}(.data, ..., .preserve = FALSE)

\method{slice}{SingleCellExperimentLogged}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{A ExperimentLogged object}

\item{...}{Row selection expressions}

\item{.preserve}{If TRUE, preserves the grouping structure of the data}
}
\description{
Slice rows from a ExperimentLogged object
}
