% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_operations.R
\name{colData<-,SummarizedExperimentLogged,DataFrame-method}
\alias{colData<-,SummarizedExperimentLogged,DataFrame-method}
\alias{colData<-,SummarizedExperimentLogged,DFrame-method}
\title{Assign column to colData using \verb{colData<-} for SummarizedExperimentLogged}
\usage{
\S4method{colData}{SummarizedExperimentLogged,DataFrame}(x) <- value

\S4method{colData}{SummarizedExperimentLogged,DFrame}(x) <- value
}
\description{
Assign column to colData using \verb{colData<-} for SummarizedExperimentLogged
}
