% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{extract,SummarizedExperimentLogged-method}
\alias{extract,SummarizedExperimentLogged-method}
\title{Extract values from a column into multiple columns in a SummarizedExperimentLogged object}
\usage{
\S4method{extract}{SummarizedExperimentLogged}(
  data,
  col,
  into,
  regex = "([[:alnum:]]+)",
  remove = TRUE,
  convert = FALSE,
  ...
)
}
\arguments{
\item{col}{Column to extract from}

\item{into}{Names of new variables to create}

\item{regex}{A regular expression to extract values}

\item{remove}{If TRUE, remove input column from output}

\item{convert}{If TRUE, runs type.convert() on each new column}

\item{...}{Additional arguments passed to tidyr::extract}

\item{.data}{A SummarizedExperimentLogged object}
}
\description{
Extract values from a column into multiple columns in a SummarizedExperimentLogged object
}
