% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{filter.SummarizedExperimentLogged}
\alias{filter.SummarizedExperimentLogged}
\alias{filter.SingleCellExperimentLogged}
\title{Filter rows and columns of a ExperimentLogged object}
\usage{
\method{filter}{SummarizedExperimentLogged}(.data, ...)

\method{filter}{SingleCellExperimentLogged}(.data, ...)
}
\arguments{
\item{.data}{A ExperimentLogged object}

\item{...}{Logical expressions used for filtering}
}
\description{
Filter rows and columns of a ExperimentLogged object
}
