% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{se_filter}
\alias{se_filter}
\alias{filter,SummarizedExperimentLogged-method}
\title{Filter rows and columns of a SummarizedExperimentLogged object}
\usage{
se_filter(.data, ...)

\S4method{filter}{SummarizedExperimentLogged}(.data, ..., .by = NULL, .preserve = FALSE)
}
\arguments{
\item{.data}{A SummarizedExperimentLogged object}

\item{...}{Logical expressions used for filtering}
}
\description{
Filter rows and columns of a SummarizedExperimentLogged object
}
