% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{mutate.SummarizedExperimentLogged}
\alias{mutate.SummarizedExperimentLogged}
\alias{mutate.SingleCellExperimentLogged}
\title{Modify columns of a ExperimentLogged object}
\usage{
\method{mutate}{SummarizedExperimentLogged}(.data, ...)

\method{mutate}{SingleCellExperimentLogged}(.data, ...)
}
\arguments{
\item{.data}{A ExperimentLogged object}

\item{...}{Name-value pairs of expressions used to modify columns}
}
\description{
Modify columns of a ExperimentLogged object
}
