% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{se_mutate}
\alias{se_mutate}
\alias{mutate,SummarizedExperimentLogged-method}
\title{Modify columns of a SummarizedExperimentLogged object}
\usage{
se_mutate(.data, ...)

\S4method{mutate}{SummarizedExperimentLogged}(.data, ...)
}
\arguments{
\item{.data}{A SummarizedExperimentLogged object}

\item{...}{Name-value pairs of expressions used to modify columns}
}
\description{
Modify columns of a SummarizedExperimentLogged object
}
