% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{select.SummarizedExperimentLogged}
\alias{select.SummarizedExperimentLogged}
\alias{select.SingleCellExperimentLogged}
\title{Filter columns of a ExperimentLogged object}
\usage{
\method{select}{SummarizedExperimentLogged}(.data, ...)

\method{select}{SingleCellExperimentLogged}(.data, ...)
}
\arguments{
\item{.data}{A ExperimentLogged object}

\item{...}{Name of columns to select or deselect}
}
\description{
Filter columns of a ExperimentLogged object
}
