% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{select,SummarizedExperimentLogged-method}
\alias{select,SummarizedExperimentLogged-method}
\title{Filter columns of a SummarizedExperimentLogged object}
\usage{
\S4method{select}{SummarizedExperimentLogged}(.data, ...)
}
\arguments{
\item{.data}{A SummarizedExperimentLogged object}

\item{...}{Name of columns to select or deselect}
}
\description{
Filter columns of a SummarizedExperimentLogged object
}
