% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{se_select}
\alias{se_select}
\alias{select,SummarizedExperimentLogged-method}
\title{Filter columns of a SummarizedExperimentLogged object}
\usage{
se_select(.data, ...)

\S4method{select}{SummarizedExperimentLogged}(.data, ...)
}
\arguments{
\item{.data}{A SummarizedExperimentLogged object}

\item{...}{Logical expressions used for filtering}
}
\description{
Filter columns of a SummarizedExperimentLogged object
}
