% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{slice}
\alias{slice}
\title{Slice rows from a SummarizedExperimentLogged object}
\usage{
slice(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{A SummarizedExperimentLogged object}

\item{...}{Row selection expressions}

\item{.preserve}{If TRUE, preserves the grouping structure of the data}
}
\description{
Slice rows from a SummarizedExperimentLogged object
}
