% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{slice,SummarizedExperimentLogged-method}
\alias{slice,SummarizedExperimentLogged-method}
\title{Slice rows from a SummarizedExperimentLogged object}
\usage{
\S4method{slice}{SummarizedExperimentLogged}(.data, ..., .by = NULL, .preserve = FALSE)
}
\description{
Slice rows from a SummarizedExperimentLogged object
}
