% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_operations.R
\name{[,SummarizedExperimentLogged,ANY,ANY,ANY-method}
\alias{[,SummarizedExperimentLogged,ANY,ANY,ANY-method}
\alias{[,SingleCellExperimentLogged,ANY,ANY,ANY-method}
\title{Subset a ExperimentLogged object}
\usage{
\S4method{[}{SummarizedExperimentLogged,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{SingleCellExperimentLogged,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A ExperimentLogged object}

\item{i, j, ...}{Indices for subsetting}
}
\description{
Subset a ExperimentLogged object
}
\examples{
# Create a logged SummarizedExperiment
if (requireNamespace("tidySummarizedExperiment", quietly = TRUE)) {
  se <- tidySummarizedExperiment::pasilla
  se_logged <- log_start(se)
  
  se_logged <- se_logged[1:10,]
}
}
