% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{[,SummarizedExperimentLogged,ANY,ANY,ANY-method}
\alias{[,SummarizedExperimentLogged,ANY,ANY,ANY-method}
\title{Subset a SummarizedExperimentLogged object}
\usage{
\S4method{[}{SummarizedExperimentLogged,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A SummarizedExperimentLogged object}

\item{i, j, ...}{Indices for subsetting}
}
\description{
Subset a SummarizedExperimentLogged object
}
