% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SummarizedExperimentLogged-class}
\alias{SummarizedExperimentLogged-class}
\title{SummarizedExperimentLogged class}
\description{
A class extending SummarizedExperiment to include logging capabilities.
This class tracks operations performed on the object and displays a
log when the object is printed.
}
\section{Slots}{

\describe{
\item{\code{log_history}}{Character vector storing the history of operations}
}}

